Modify migration to only attempt to drop admins table if it exists.

+++ db/migrate/20150317064814_devise_create_admins.rb
@@ -1,6 +1,6 @@
 class DeviseCreateAdmins < ActiveRecord::Migration
   def change
-    drop_table :admins
+    drop_table :admins if table_exists? :admins
     create_table(:admins) do |t|
       ## Database authenticatable
       t.string :email,              null: false, default: ""
